app CRUD 第三次提交

+++ admin-console/src/main/java/io/information/modules/app/entity/InCardVote.java
@@ -26,7 +26,7 @@ import java.time.LocalDateTime;
             /**
             * 帖子id
             */
-    private Integer cId;
+    private Long cId;
 
             /**
             * 投票选项信息（逗号分隔）
